refactor
components for single responsibility

extract DOM helpers into dom.ts and split editing/pinning concern

Add inline-editing coverage for keyboard, checkbox, and select editors.
