+++ src/components/DataTableBody.tsx
@@ -2,7 +2,8 @@ import * as React from 'react';
 import Body from './TableBody';
 import Row from './TableRow';
 import NoData from './NoDataWrapper';
-import { prop, isEmpty, flipElement } from '../util';
+import { prop, isEmpty } from '../util';
+import { flipElement } from '../dom';
 import type { TableRow, RowState } from '../types';
 import { useRowContext } from '../context/RowContext';
 import useIsomorphicLayoutEffect from '../hooks/useIsomorphicLayoutEffect';
